pyproject.toml
[project]
name = "evidenceplatform"
version = "0.1.0"
description = "Evidence archiving and browsing platform"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.14.3",
    "cryptography>=46.0.5",
    "har2warc==1.0.4",
    "ijson==3.5.0",
    "mysql-connector-python==9.6.0",
    "numpy==2.4.4",
    "opentimestamps-client==0.7.2",
    "pillow>=12.1.1",
    "playwright==1.58.0",
    "psutil==6.1.1",
    "pyasn1>=0.6.2",
    "pyautogui==0.9.54",
    "pydantic==2.12.5",
    "pygetwindow==0.0.9",
    "pyinstaller==6.19.0",
    "pyperclip==1.11.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "requests>=2.32.4",
    "rfc3161ng==2.1.3",
    "safety>=3.5.2",
    "tqdm==4.67.3",
    "zstandard==0.25.0",
    "fastapi>=0.123.0",
    "uvicorn>=0.38.0",
    "websockets>=14.0",
    "argon2-cffi>=25.1.0",
    "tzlocal>=5.3.1",
    "pytz>=2025.2",
    # pinned: 4.13.0.92 lacks Linux wheels, waiting for fix https://pypi.org/project/opencv-python-headless/#history
    "opencv-python-headless==4.13.0.90",
    "ppdeep>=20260221",
    "starlette>=0.50.0,<1.0.0",
    # capped: starlette 1.0 is a major version bump; review middleware API changes before upgrading,
    "toga>=0.5.3",
    "warcio>=1.8.1",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.24",
    "pyotp>=2.9.0",
    "qrcode[pil]>=8.0",
    "slowapi>=0.1.9",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.deptry]
# CLI tools (pyinstaller, safety) and runtime-only deps (websockets used by starlette,
# opentimestamps-client called via subprocess) are not imported directly.
known_first_party = [
    "browsing_platform",
    "archiver",
    "extractors",
    "db_loaders",
    "utils",
    "root_anchor",
]

[tool.deptry.per_rule_ignores]
DEP002 = ["opentimestamps-client", "pyinstaller", "safety", "websockets"]

[tool.deptry.package_module_name_map]
beautifulsoup4 = "bs4"
python-dotenv = "dotenv"
opencv-python-headless = "cv2"
pillow = "PIL"
python-dateutil = "dateutil"
mysql-connector-python = "mysql"
"argon2-cffi" = "argon2"

[tool.bandit]
# B608 (hardcoded SQL expressions) is skipped globally because all f-string SQL
# queries in this codebase are either:
#   (a) parameterized with %(key)s placeholders only (no user input in the string), or
#   (b) constructed with column/table names from a hardcoded allowlist (ENTITY_TAG_TABLES,
#       ALLOWED_COLUMNS) validated by sanitize_column() before use.
# See security-scans.md for full analysis.
skips = ["B608"]
